+++ docs/source/release-notes.rst
@@ -7,9 +7,11 @@ Release 0.2.1
 
 * Class ``LoggerFactory``. Logger factory for configuration of the Python logging framework.
 
-* Singleton decorator utilizes wrapt behaving much nicer.
-See https://www.youtube.com/watch?v=W7Rv-km3ZuA&spfreload=10 for Graham Dumpleton's talk
-on the workings of wrapt.
+* The ``fuzzer`` module uses logging.
+
+* Singleton decorator behaves much nicer since using ``wrapt``.
+  See `Graham Dumpleton's talk <https://www.youtube.com/watch?v=W7Rv-km3ZuA&spfreload=10>`_
+  on the workings of wrapt.
 
 
 Release 0.2.0